#118 - add default values to review edition
