+++ client/src/api/reviews.ts
@@ -1,0 +1,32 @@
+import { apiRoutes } from '../routes/apiRoutes';
+import {
+  useDelete,
+  useFetch,
+  usePost,
+  useUpdate,
+} from '../utils/reactQuery.utils';
+import { pathToUrl } from '../utils/router.utils';
+import { Review, ReviewData } from '../utils/types';
+
+export const useGetReviews = (productId?: string) =>
+  useFetch<ReviewData>(
+    `${apiRoutes.getReviews}?${productId && `product_id=${productId}&`}sort=created_at:desc`
+  );
+
+export const useGetReview = (reviewId: string) =>
+  useFetch<Review>(pathToUrl(apiRoutes.getReview, { reviewId }));
+
+export const useAddReview = (productId: string) =>
+  usePost(pathToUrl(apiRoutes.addReview, { productId }), undefined, {
+    withCredentials: true,
+  });
+
+export const useUpdateReview = (reviewId: string) =>
+  useUpdate(pathToUrl(apiRoutes.updateReview, { reviewId }), undefined, {
+    withCredentials: true,
+  });
+
+export const useDeleteReview = () =>
+  useDelete(apiRoutes.deleteReview, undefined, {
+    withCredentials: true,
+  });
